tests/wpt/tests/html/interaction/focus/focusgroup/tentative/forward-navigation/horizontal/rtl-inline-only.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - RTL with inline-only axis respects reversed directional input.</title>
<meta name="assert" content="In an RTL focusgroup restricted to inline axis, Left moves forward and Right moves backward. Up/Down should not move focus since the block axis is not supported.">
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../../resources/focusgroup-utils.js"></script>

<div dir="rtl" focusgroup="toolbar inline">
  <span id=item1 tabindex=0>item1</span>
  <span id=item2 tabindex=0>item2</span>
  <span id=item3 tabindex=0>item3</span>
</div>

<script>

  promise_test(async t => {
    var item1 = document.getElementById("item1");
    var item2 = document.getElementById("item2");

    // Left is forward inline in RTL.
    await focusAndSendDirectionalInput(item1, kLeft);
    assert_equals(document.activeElement, item2,
      "Left should move forward inline in RTL inline-only focusgroup");
  }, "Left moves forward in RTL inline-only focusgroup.");

  promise_test(async t => {
    var item1 = document.getElementById("item1");
    var item2 = document.getElementById("item2");

    // Right is backward inline in RTL.
    await focusAndSendDirectionalInput(item2, kRight);
    assert_equals(document.activeElement, item1,
      "Right should move backward inline in RTL inline-only focusgroup");
  }, "Right moves backward in RTL inline-only focusgroup.");

  promise_test(async t => {
    var item1 = document.getElementById("item1");

    // Block axis keys should not move focus in an inline-only focusgroup.
    await focusAndSendDirectionalInput(item1, kDown);
    assert_equals(document.activeElement, item1,
      "Down should not move focus in inline-only focusgroup");

    await focusAndSendDirectionalInput(item1, kUp);
    assert_equals(document.activeElement, item1,
      "Up should not move focus in inline-only focusgroup");
  }, "Down and Up do not move focus in RTL inline-only focusgroup.");

</script>
